requirements.txt
streamlit~=1.8.1
"pip install streamlit"

requests~=2.27.1
"pip install requests"

Pillow~=9.1.0
"pip install Pillow"

torch~=1.11.0+cu113
torchvision~=0.12.0+cu113
"pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu113"

numpy~=1.22.3
"pip install numpy"

